requirements.txt
fastapi
uvicorn
requests
python-dotenv
openai
psycopg2-binary
beautifulsoup4
requests
sqlalchemy>=2.0
alembic
pydantic
beautifulsoup4
selenium
webdriver-manager
lxml